--- jaxfg/__init__.py
@@ -1,12 +1,0 @@
-from . import core, experimental, geometry, hints, noises, solvers, sparse, utils
-
-__all__ = [
-    "core",
-    "experimental",
-    "geometry",
-    "hints",
-    "noises",
-    "solvers",
-    "sparse",
-    "utils",
-]
