--- jaxfg/core/__init__.py
@@ -1,16 +1,0 @@
-from ._factor_base import FactorBase
-from ._factor_stack import FactorStack
-from ._stacked_factor_graph import StackedFactorGraph
-from ._storage_layout import StorageLayout
-from ._variable_assignments import VariableAssignments
-from ._variables import RealVectorVariable, VariableBase
-
-__all__ = [
-    "FactorStack",
-    "FactorBase",
-    "StackedFactorGraph",
-    "StorageLayout",
-    "VariableAssignments",
-    "RealVectorVariable",
-    "VariableBase",
-]
