--- jaxfg/experimental/__init__.py
@@ -1,3 +1,0 @@
-from ._sparse_covariance import SparseCovariance
-
-__all__ = ["SparseCovariance"]
