--- jaxfg/geometry/__init__.py
@@ -1,18 +1,0 @@
-from ._factors import BetweenFactor, PriorFactor
-from ._lie_variables import (
-    LieVariableBase,
-    SE2Variable,
-    SE3Variable,
-    SO2Variable,
-    SO3Variable,
-)
-
-__all__ = [
-    "BetweenFactor",
-    "PriorFactor",
-    "LieVariableBase",
-    "SE2Variable",
-    "SE3Variable",
-    "SO2Variable",
-    "SO3Variable",
-]
