--- jaxfg/hints/__init__.py
@@ -1,9 +1,0 @@
-from ._aliases import Array, LocalVariableValue, Pytree, Scalar, VariableValue
-
-__all__ = [
-    "Array",
-    "LocalVariableValue",
-    "Pytree",
-    "Scalar",
-    "VariableValue",
-]
