--- jaxfg/sparse/__init__.py
@@ -1,16 +1,0 @@
-from ._linear_solve import (
-    CholmodSolver,
-    ConjugateGradientSolver,
-    InexactStepConjugateGradientSolver,
-    LinearSubproblemSolverBase,
-)
-from ._sparse_matrix import SparseCooCoordinates, SparseCooMatrix
-
-__all__ = [
-    "CholmodSolver",
-    "ConjugateGradientSolver",
-    "InexactStepConjugateGradientSolver",
-    "LinearSubproblemSolverBase",
-    "SparseCooCoordinates",
-    "SparseCooMatrix",
-]
