--- scripts_v2/ok.py
@@ -1,7 +1,0 @@
-import pinocchio as pin
-import jaxlie
-import jaxlie.manifold
-import numpy as np
-T = pin.SE3.Random()
-pin.log6(T)
-breakpoint()
